Edit feature

+++ db/migrate/20171106152434_create_ingredients.rb
@@ -2,7 +2,7 @@ class CreateIngredients < ActiveRecord::Migration[5.1]
   def change
     create_table :ingredients do |t|
       t.string :item, null: false
-      t.integer :amount, null: false
+      t.float :amount, null: false
       t.string :measurement, null:false
       t.references :recipe, foreign_key: true, null: false
